hugging-face/intro-hf-inference-api/pyproject.toml
[tool.poetry]
name = "intro-hf-inference-api"
version = "0.1.0"
description = ""
authors = ["Helton Carlos de Souza <heltoncarlossouza@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
huggingface-hub = "^0.19.1"
python-dotenv = "^1.0.0"
pillow = "^10.1.0"
numpy = "^1.26.2"
tabulate = "^0.9.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
poethepoet = "^0.24.3"

[tool.poe.tasks.format]
shell = "black . && isort ."

[tool.poe.tasks.requirements]
shell = "poetry export --without-hashes --format=requirements.txt > requirements.txt"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
